requirements.txt
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
dominate==2.9.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-Testing==0.8.1
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.29.1
MarkupSafe==2.1.5
msgpack==1.0.8
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
PyMySQL==1.1.1
PySocks==1.7.1
pytest==8.2.2
pytest-sugar==1.0.0
python-dotenv==1.0.1
pyzmq==26.0.3
requests==2.32.3
selenium==4.22.0
setuptools==70.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
termcolor==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
visitor==0.1.3
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
